Refactoring GroupMgt
