+++ GroupMgt/tool.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+# coding: utf-8 -*-
+#
+# Author: zaraki673 & pipiche38
+#
+
+
+GROUPS_CONFIG_FILENAME = "ZigateGroupsConfig"
+TIMEOUT = 12
+MAX_CYCLE = 3
